refactor: move API Keys into the Integrations page as a tab
Drop the separate 'API Keys' side-nav entry. The key management UI now lives
as an 'API Keys' tab on Integrations (deep-linkable via ?tab=api-keys), so
the sidebar stays unchanged.
